Refactor to some abstractions
* Entities are things in...
* World, which manages Entity state each tick, which is controlled by...
* Game, which also sets up some canvas stuff.

It looks like a design is emerging, where responsibility for managing
Entity state and drawing is pushed as close to the Entity as possible.

It might be sensible that the game does the Rendering, or it is provided
some renderer to do so. That would better separate the tick, which is
Game stuff and won't change significantly (save managing framecount),
from the render, which is deep inside each entity.